gh-111506: Rename _Py_SetRefcnt() to Py_SET_REFCNT()
Export Py_SET_REFCNT() opaque function as "Py_SET_REFCNT" symbol in
the stable ABI.

Add  Objects/object_abi.c file and _Py_STABLE_ABI_IMPL macro.
