+++ Lib/test/test_stable_abi_ctypes.py
@@ -870,6 +870,7 @@ SYMBOL_NAMES = (
     "Py_NewRef",
     "Py_ReprEnter",
     "Py_ReprLeave",
+    "Py_SET_REFCNT",
     "Py_SetPath",
     "Py_SetProgramName",
     "Py_SetPythonHome",
@@ -906,7 +907,6 @@ SYMBOL_NAMES = (
     "_Py_IncRef",
     "_Py_NoneStruct",
     "_Py_NotImplementedStruct",
-    "_Py_SetRefcnt",
     "_Py_SwappedOp",
     "_Py_TrueStruct",
     "_Py_VaBuildValue_SizeT",
