+++ Objects/object.c
@@ -2886,55 +2886,3 @@ PyObject_GET_WEAKREFS_LISTPTR(PyObject *op)
 {
     return _PyObject_GET_WEAKREFS_LISTPTR(op);
 }
-
-
-#undef Py_NewRef
-#undef Py_XNewRef
-
-// Export Py_NewRef() and Py_XNewRef() as regular functions for the stable ABI.
-PyObject*
-Py_NewRef(PyObject *obj)
-{
-    return _Py_NewRef(obj);
-}
-
-PyObject*
-Py_XNewRef(PyObject *obj)
-{
-    return _Py_XNewRef(obj);
-}
-
-#undef Py_Is
-#undef Py_IsNone
-#undef Py_IsTrue
-#undef Py_IsFalse
-
-// Export Py_Is(), Py_IsNone(), Py_IsTrue(), Py_IsFalse() as regular functions
-// for the stable ABI.
-int Py_Is(PyObject *x, PyObject *y)
-{
-    return (x == y);
-}
-
-int Py_IsNone(PyObject *x)
-{
-    return Py_Is(x, Py_None);
-}
-
-int Py_IsTrue(PyObject *x)
-{
-    return Py_Is(x, Py_True);
-}
-
-int Py_IsFalse(PyObject *x)
-{
-    return Py_Is(x, Py_False);
-}
-
-
-// Py_SET_REFCNT() implementation for stable ABI
-void
-_Py_SetRefcnt(PyObject *ob, Py_ssize_t refcnt)
-{
-    Py_SET_REFCNT(ob, refcnt);
-}
