+++ PC/python3dll.c
@@ -19,7 +19,6 @@ EXPORT_FUNC(_Py_Dealloc)
 EXPORT_FUNC(_Py_DecRef)
 EXPORT_FUNC(_Py_IncRef)
 EXPORT_FUNC(_Py_NegativeRefcount)
-EXPORT_FUNC(_Py_SetRefcnt)
 EXPORT_FUNC(_Py_VaBuildValue_SizeT)
 EXPORT_FUNC(_PyArg_Parse_SizeT)
 EXPORT_FUNC(_PyArg_ParseTuple_SizeT)
@@ -81,6 +80,7 @@ EXPORT_FUNC(Py_NewInterpreter)
 EXPORT_FUNC(Py_NewRef)
 EXPORT_FUNC(Py_ReprEnter)
 EXPORT_FUNC(Py_ReprLeave)
+EXPORT_FUNC(Py_SET_REFCNT)
 EXPORT_FUNC(Py_SetPath)
 EXPORT_FUNC(Py_SetProgramName)
 EXPORT_FUNC(Py_SetPythonHome)
